requirements.txt
pyserial
psutil
pillow
keyboard
pyyaml
